remaster project-files-explorer.component
